Finish viewProductionModal and main.js for view production log
main.js - Fix vMat2 not displaying material name when modal is viewed, input tag was spelled wrong.

Removed button at the top of the viewProducitonModal because it closed the modal but left the page gray out.
The cancel button at the bottom of the modal works as intended.
Removed add log button because it is not needed.
Filled out the rest of the form except for the daily numbers.