fixing viewProductionModal
When modal was loading it was failing to add data to the daily usage elements on the page.  Parsed the values in getLog function and the seemed to work. Values were added to the elements for daily production but they were incorrect.

The reason they were incorrect was my query to the db was throwing an error and returning nothing.  Removed the t.* from my query and this correct the issue for the query and started returning the correct data.

The wrong values were still being displayed in the daily production and I found that I was using the wrong response. value for the previous log.

Added .toFixed(3) while adding values to limit decimal places to 3.