Finsihed QA Reject submission
added new javascript qaRejects.js which handles the form submission of data and refreshing of last 4 weeks table.

added util.php for the show messages
modified main.js fetchLast4weeks so that is was global and could be triggered from inside of qaRejects.js

added PDO Transaction to fetch_l4w_logs.php to handed the insert and update.

add isset for the qaReject form data to the productionActions.php