still working on production log insert
Decided to rewrite the switch statement that is triggered base on the production run status because it was too unwieldy.

I wrote a function that handles the fetching of the material data from previous log for in progress and end of production run.

In progress is return the right log but the data  isn't being added to daily usage properly.

End of run is not returning a record and throwing errors.

Need to make sure Start is working properly.