fleshing out Quality namespace
Also noticed that when removing copper pins forgot to create a transaction for them.

Noticed that I am using material run usage for each hopper to subtract material and add transactions.  Added dailyHopUsed# to the matlog table to stores the dailyUsage per hopper, but now have to refactor some code.